requirements.txt
python-telegram-bot
pysocks